multi-agent/requirements.txt
smolagents[litellm]
plotly
geopandas
shapely
kaleido
Pillow
certifi